third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/cross-origin-opener-policy/reporting/access-reporting/report-to-both_coop-ro.https-expected.txt

This is a testharness.js-based test.
[FAIL] Access from opener
  assert_equals: expected (string) "coop" but got (undefined) undefined
[FAIL] Access from openee
  assert_equals: expected (string) "coop" but got (undefined) undefined
Harness: the test ran to completion.
